# apps/api/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DSM — API de administración del catálogo (US-001)
  version: 1.0.0
  description: >-
    Panel del dueño: categorías y productos, máquina de estado y RBAC admin.
    Errores en envelope RFC 7807 (`application/problem+json`, `type` `dsm:catalog/*`).
  contact:
    name: DSM
    url: https://example.com
servers:
  - url: /v1
    description: Base URL versionada
...
security:
  - adminBearer: []
paths:
...
  /admin/categories:
    post:
      operationId: createCategory
      tags: [admin-categories]
      summary: Crear categoría (AC-1)
      description: Crea una categoría; el slug se deriva del nombre (único).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCategory' }
      responses:
        '201': { description: Creada, content: { application/json: { schema: { $ref: '#/components/schemas/Category' } } } }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
    get:
      operationId: listCategories
      tags: [admin-categories]
      summary: Listar categorías (AC-1)
      description: Devuelve todas las categorías.
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Category' } } } } }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /admin/categories/{id}:
    patch:
      operationId: updateCategory
      tags: [admin-categories]
      summary: Editar categoría (AC-1)
      description: Actualiza el nombre/parent de una categoría.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateCategory' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Category' } } } }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
...
  /categories/{slug}:
    get:
      operationId: storefrontGetCategory
      tags: [storefront-categories]
      summary: Detalle público de una categoría (US-002 AC-1/AC-2/AC-9)
      description: >-
        Ruta PÚBLICA SIN auth. Trae `parent` para el breadcrumb y `children`
        para los subrubros. Slug inexistente → 404 (nunca un 200 vacío, que
        sería una página fantasma indexable). Único endpoint que emite el evento
        `category.viewed` (D4, insumo del panel US-016).
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$' }
      responses:
        '200':
          description: Categoría con su padre y sus subrubros
          headers:
            Cache-Control:
              schema: { type: string }
              description: 'public, max-age=60, stale-while-revalidate=30'
          content: { application/json: { schema: { $ref: '#/components/schemas/StorefrontCategory' } } }
        '404': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Carrito del invitado con precios vigentes y disponibilidad (US-007 AC-4/AC-6/AC-7/AC-9)
      description: >-
        Ruta PÚBLICA SIN auth. Operación **segura**: no crea carrito, no emite
        cookie y no desliza la ventana de retención. Sin cookie `dsm_cart`, con una
        cookie huérfana o con la fila vencida devuelve el carrito VACÍO con 200
        (AC-7) — `id: null`, `items: []`, contadores en 0 y `updated_at: null`; la
        fila vencida se purga en el acto. Cada lectura recalcula precio y
        disponibilidad contra `products` (AC-6/AC-9), así que una línea
        despublicada o sin stock suficiente aparece marcada y fuera del total, no
        borrada. `Cache-Control: no-store` en toda la superficie.
      security: []
      responses:
        '200':
          description: Carrito del invitado (vacío si no hay cookie)
          headers:
            Cache-Control:
              schema: { type: string }
              description: 'Siempre no-store (§7.1): sin esto AC-9 se cae en el edge.'
          content: { application/json: { schema: { $ref: '#/components/schemas/CartEnvelope' } } }
        '429': { $ref: '#/components/responses/CartRateLimited' }
  /cart/items/{slug}:
    put:
      operationId: setCartItem
      tags: [cart]
      summary: Fija la cantidad ABSOLUTA de un producto en el carrito (AC-1/AC-2/AC-5/AC-10)
      description: >-
        Ruta PÚBLICA SIN auth. El cuerpo fija la cantidad, no la suma: la operación
        es naturalmente idempotente (api-standards §10.5) y por eso no hay
        `Idempotency-Key`. Crea la línea si no existe, y crea el carrito + emite las
        cookies si no había. Orden de validación producto → stock → carrito, así que
        un rechazo NO escribe nada. `quantity` mayor al stock → 409
        `dsm:cart/insufficient-stock` con `available_quantity` (AC-5); el stock NO
        se reserva ni se descuenta (AC-8). Un producto `draft`, `archived` o
        inexistente devuelve el MISMO 404 (AC-10). Superar `CART_MAX_ITEMS` líneas
        → 409 `dsm:cart/too-many-items` con `max_items`.
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          description: 'Identificador público del producto (US-002/US-003 no exponen id).'
          schema: { type: string, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$' }
        - $ref: '#/components/parameters/CartCsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetCartItemRequest' }
      responses:
        '200':
          description: Carrito completo recalculado
          headers:
            Set-Cookie: { $ref: '#/components/headers/CartCookies' }
            Cache-Control:
              schema: { type: string }
              description: 'Siempre no-store.'
          content: { application/json: { schema: { $ref: '#/components/schemas/CartEnvelope' } } }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/CartRateLimited' }
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Quita una línea del carrito (AC-3)
      description: >-
        Ruta PÚBLICA SIN auth. Idempotente: quitar algo que no está devuelve el
        carrito igual con 200 y sin error. El producto se resuelve SIN filtrar por
        estado, para que una línea despublicada (AC-6) se pueda sacar. No crea
        carrito: sin cookie devuelve el vacío.
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$' }
        - $ref: '#/components/parameters/CartCsrfToken'
      responses:
        '200':
          description: Carrito recalculado (vacío si no había carrito)
          headers:
            Set-Cookie: { $ref: '#/components/headers/CartCookies' }
            Cache-Control:
              schema: { type: string }
              description: 'Siempre no-store.'
          content: { application/json: { schema: { $ref: '#/components/schemas/CartEnvelope' } } }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/CartRateLimited' }
...
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT con claim role=admin (seam ADR-0009)
    sessionCookie:
      type: apiKey
      in: cookie
      name: dsm_access
      description: >-
        Access token de cliente, corto y HttpOnly (US-014). La API NO lo acepta
        por header `Authorization`: aceptarlo permitiría a un XSS usar un token
        robado y anularía la protección de HttpOnly.
  headers:
...
    CartCookies:
      description: >-
        Dos cookies del carrito del invitado (§7.4): dsm_cart (HttpOnly,
        SameSite=Lax, Path=/, Max-Age = CART_TTL_DAYS × 86400) con el token opaco, y
        dsm_cart_csrf (SIN HttpOnly — el frontend la lee para armar X-CSRF-Token),
        que es HMAC(JWT_SECRET, token). Secure sale de AUTH_COOKIE_SECURE. El token
        NUNCA viaja en el cuerpo, y la cookie y la fila `carts.expires_at` vencen
        juntas porque derivan del mismo CART_TTL_DAYS.
      schema: { type: string }
  parameters:
    CartCsrfToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: >-
        Double-submit firmado del carrito (§7.5): el valor de la cookie
        dsm_cart_csrf. Se exige junto con un Origin de la allowlist, cuya ausencia
        también se rechaza con 403. Es obligatorio CUANDO la petición trae la cookie
        dsm_cart; la primera escritura de un cliente nuevo no la trae y pasa, porque
        todavía no hay carrito que secuestrar.
      schema: { type: string }
...
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    CartRateLimited:
      description: >-
        Presupuesto del throttler `cart` excedido (§7.3): CART_RATE_LIMIT_MAX
        lecturas y CART_WRITE_RATE_LIMIT_MAX escrituras por minuto y por IP,
        contadas por endpoint. Incluye Retry-After y las cabeceras RateLimit-*
        (api-standards §12) y `Cache-Control: no-store` — un 429 cacheado en el edge
        convertiría el rate-limit en un DoS.
      headers:
        Retry-After: { schema: { type: integer } }
        RateLimit-Limit: { schema: { type: integer } }
        RateLimit-Remaining: { schema: { type: integer } }
        RateLimit-Reset: { schema: { type: integer } }
        Cache-Control: { schema: { type: string } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Problem:
      description: Error RFC 7807
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    CreateCategory:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        parent_id: { type: string, format: uuid, nullable: true }
    UpdateCategory:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        parent_id: { type: string, format: uuid, nullable: true }
    Category:
      type: object
      required: [id, slug, name, parent_id, created_at]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        parent_id: { type: string, format: uuid, nullable: true }
        created_at: { type: string, format: date-time }
...
    CategoryRef:
      type: object
      required: [slug, name]
      properties:
        slug: { type: string }
        name: { type: string }
...
    StorefrontCategory:
      type: object
      description: 'Categoría pública (US-002): sin id, parent_id ni timestamps.'
      required: [slug, name, parent, children]
      properties:
        slug: { type: string }
        name: { type: string }
        parent:
          type: object
          nullable: true
          description: 'Rubro padre para el breadcrumb; null si ya es raíz.'
          required: [slug, name]
          properties:
            slug: { type: string }
            name: { type: string }
        children:
          type: array
          items: { $ref: '#/components/schemas/CategoryRef' }
...
    SetCartItemRequest:
      type: object
      description: >-
        Entrada de PUT /cart/items/{slug} — UN SOLO campo a propósito. No hay
        unit_price_ars_cents, product_id ni cart_id: todo importe se deriva
        server-side del precio vigente y el carrito se identifica por la cookie.
        Con forbidNonWhitelisted, mandar un campo desconocido es 422 (no se ignora).
      required: [quantity]
      additionalProperties: false
      properties:
        quantity:
          type: integer
          minimum: 1
          maximum: 99
          description: >-
            Cantidad ABSOLUTA de unidades del producto en el carrito. El tope es
            CART_MAX_QTY_PER_LINE (99 por defecto, cota anti-DoS §7.3). No se acepta
            0: cantidad 0 no es una línea, es un DELETE.
    CartItem:
      type: object
      description: >-
        Línea del carrito (US-007). Sin product_id, cart_id, status ni stock crudo:
        el identificador público es el slug y el inventario se comunica como
        availability + max_quantity (OQ-BE-2).
      required:
        - slug
        - name
        - image_url
        - quantity
        - unit_price_ars_cents
        - currency
        - subtotal_ars_cents
        - availability
        - max_quantity
        - price_changed
      properties:
        slug: { type: string }
        name: { type: string }
        image_url:
          type: string
          nullable: true
        quantity: { type: integer, minimum: 1 }
        unit_price_ars_cents:
          type: integer
          description: 'Precio VIGENTE en centavos ARS (AC-9), no la instantánea guardada.'
        currency: { type: string, enum: [ARS] }
        subtotal_ars_cents:
          type: integer
          description: >-
            unit_price_ars_cents × quantity. Se calcula también en las líneas
            bloqueadas, aunque no entren en el total.
        availability:
          type: string
          enum: [available, insufficient_stock, unavailable]
          description: >-
            available: publicado con stock ≥ cantidad. insufficient_stock: publicado
            con stock < cantidad (AC-6). unavailable: draft o archived (AC-6). La
            línea nunca se borra sola.
        available_quantity:
          type: integer
          description: 'Presente sólo cuando la línea pide más de lo que hay (OQ-BE-2).'
        max_quantity:
          type: integer
          description: 'min(stock, CART_MAX_QTY_PER_LINE) — techo del stepper.'
        price_changed:
          type: boolean
          description: >-
            true si el precio vigente difiere de la instantánea del último toque: el
            cambio se muestra, no se aplica en silencio.
        previous_unit_price_ars_cents:
          type: integer
          description: 'Presente sólo si price_changed es true.'
    Cart:
      type: object
      required:
        - id
        - items
        - item_count
        - total_quantity
        - total_ars_cents
        - has_blocking_issues
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          nullable: true
          description: >-
            null cuando todavía no hay carrito (AC-7). Conocerlo NO da acceso: el
            acceso es la cookie dsm_cart, y no hay ninguna ruta que acepte un id de
            carrito.
        items: { type: array, items: { $ref: '#/components/schemas/CartItem' } }
        item_count: { type: integer, description: 'Líneas distintas (no unidades).' }
        total_quantity: { type: integer, description: 'Unidades, incluidas las bloqueadas.' }
        total_ars_cents:
          type: integer
          description: >-
            Suma de los subtotales de las líneas `available` SOLAMENTE (OQ-BE-4): un
            total que incluyera lo no comprable es un número que el checkout va a
            desmentir.
        has_blocking_issues:
          type: boolean
          description: >-
            true si alguna línea no es available. Es la señal con la que US-008
            impide avanzar al pago.
        updated_at:
          type: string
          format: date-time
          nullable: true
    CartEnvelope:
      type: object
      required: [cart]
      properties:
        cart: { $ref: '#/components/schemas/Cart' }
    Problem:
      type: object
      description: RFC 7807
      required: [type, title, status, detail, instance]
      properties:
        type: { type: string, example: 'dsm:catalog/conflict' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        available_quantity:
          type: integer
          description: >-
            Extension member RFC 7807 §3.2 de `dsm:cart/insufficient-stock`
            (US-007 AC-5): unidades realmente disponibles. Va como campo de primer
            nivel para que el frontend no tenga que sacarlo del detail con una regex.
        max_items:
          type: integer
          description: >-
            Extension member de `dsm:cart/too-many-items`: tope de líneas distintas
            por carrito (CART_MAX_ITEMS).
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...